creating a abstractelement to be reused in Graphene
